requirements.txt
# Core
django==4.2.9
django-allauth==0.60.1
PyJWT==2.8.0
python-dotenv==1.0.1
str2bool==1.1
django-quill-editor==0.1.40

# Misc
celery==5.3.4
redis==5.0.1
hiredis==2.2.3
pillow==11.1.0


# Tools
django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
django-debug-toolbar==4.2.0
drf-spectacular==0.26.5

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0
cryptography==42.0.2

# DB Layer
# psycopg2-binary
# mysqlclient

# ENTERPRISE Version
#  - CVS to Model
#  - Loader 
# pandas==2.2.0
django-datatables-view==1.20.0
pywebpush
reportlab
django-pwa
